bsps/arm: Fix bsp_section_data_end linker symbol
The bsp_section_data_end linker symbol indicates the end of region
REGION_DATA, which includes the sections data, data1, and rtemsrwset.
The rtemsrwset section will be empty if linker sets are not used in the
build.

When the zImage header patch was submitted, the rtemsrwset section
had a 1 byte alignment. The rtemsrwset section now has a 64 byte
alignment - this change occurred in commit 234d155e "linkersets: Revert
to zero-length arrays".

In the case that the rtemsrwset section is empty, the
bsp_section_data_end symbol will contain start address of this section,
which is rouded up to the alignment. This does not match the actual end
address of the region, which would be the end of data1.

This mismatch becomes apparent when using objcopy to convert the ELF to
a BIN image - objcopy does not pad up to the rtemsrwset section since it
is empty. As a result, the length of the BIN image does not match the
bsp_section_data_end symbol, which is used in the zImage header. Certain
image loaders will check for this and throw an error.

This change adds a conditional statement to the linker script that will
use the end of the data1 section in the case that rtemsrwset is empty.